+++ pronamic-pay-gravityforms.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * Plugin Name: Pronamic Pay Gravity Forms Add-On
+ * Plugin URI: https://www.pronamic.eu/plugins/pronamic-pay-gravityforms/
+ * Description: Extend the Pronamic Pay plugin with Gravity Forms support to receive payments through a variety of payment providers.
+ *
+ * Version: 4.1.0
+ * Requires at least: 4.7
+ *
+ * Author: Pronamic
+ * Author URI: https://www.pronamic.eu/
+ *
+ * Text Domain: pronamic-pay-gravityforms
+ * Domain Path: /languages/
+ *
+ * License: GPL-3.0-or-later
+ *
+ * Depends: wp-pay/core
+ *
+ * GitHub URI: https://github.com/pronamic/wp-pronamic-pay-gravityforms
+ *
+ * @author    Pronamic <info@pronamic.eu>
+ * @copyright 2005-2022 Pronamic
+ * @license   GPL-3.0-or-later
+ * @package   Pronamic\WordPress\Pay\Extensions\GravityForms
+ */
+
+add_filter(
+	'pronamic_pay_plugin_integrations',
+	function ( $integrations ) {
+		foreach ( $integrations as $integration ) {
+			if ( $integration instanceof \Pronamic\WordPress\Pay\Extensions\GravityForms\Extension ) {
+				return $integrations;
+			}
+		}
+
+		$integrations[] = new \Pronamic\WordPress\Pay\Extensions\GravityForms\Extension();
+
+		return $integrations;
+	}
+);