+++ src/Util.php
@@ -34,25 +34,25 @@ class Util {
 	 */
 	public static function get_detected_field_id( $field_name, $form, $entry ) {
 		// Field types with optional input ID suffix as key.
-		$types = array(
-			'name'    => array(
+		$types = [
+			'name'    => [
 				2 => 'prefix_name',
 				3 => 'first_name',
 				4 => 'middle_name',
 				6 => 'last_name',
 				8 => 'suffix_name',
-			),
-			'address' => array(
+			],
+			'address' => [
 				1 => 'address1',
 				2 => 'address2',
 				3 => 'city',
 				4 => 'state',
 				5 => 'zip',
 				6 => 'country',
-			),
-			'phone'   => array( 'telephone_number' ),
-			'email'   => array( 'email' ),
-		);
+			],
+			'phone'   => [ 'telephone_number' ],
+			'email'   => [ 'email' ],
+		];
 
 		// Determine type and input ID suffix (if applicable).
 		$input_type   = null;
@@ -68,7 +68,7 @@ class Util {
 			$input_type = $type;
 
 			// Input ID suffix.
-			if ( \in_array( $type, array( 'name', 'address' ), true ) ) {
+			if ( \in_array( $type, [ 'name', 'address' ], true ) ) {
 				$input_suffix = $search;
 			}
 
@@ -83,7 +83,7 @@ class Util {
 			}
 
 			// Check field visibility.
-			if ( RGFormsModel::is_field_hidden( $form, $field, array(), $entry ) ) {
+			if ( RGFormsModel::is_field_hidden( $form, $field, [], $entry ) ) {
 				continue;
 			}